requirements.txt
autopep8
celery
cryptography
coreapi
Django
django-admin-sortable2
django-celery-beat
django_celery_results
django-cleanup
django-cors-headers
django-htmlmin
django-health-check
django-jazzmin
django-pictures
django-post_office
django-rest-swagger
django-simple-history
django-tinymce
djangorestframework
djangorestframework_simplejwt
drf_spectacular
filetype
gunicorn
Flask-SQLAlchemy
Markdown
mysqlclient
orjson
pdfkit
Pillow
psutil
python-dotenv
python-telegram-bot
python-dateutil
pyyaml
requests
requests-toolbelt
sentry-sdk
twilio
Unidecode
urllib3
virtualenv
wheel
